% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-ensemblVEP.R
\name{ensemblVEP,character-method}
\alias{ensemblVEP,character-method}
\title{basic method}
\usage{
\S4method{ensemblVEP}{character}(file, param = VEPFlags(), ..., verbose = FALSE)
}
\arguments{
\item{file}{A \code{character} specifying the full path to the file,
including the file name.}

\item{param}{An instance of \code{VEPFlags} specifying runtime options.}

\item{\dots}{Additional arguments passed to methods.}

\item{verbose}{logical(1) should system call to vep be printed}
}
\description{
basic method
}
