\name{runtimeOptions}

\alias{runtimeOptions}

\title{VEPParam runtime options}

\description{
  Runtime options for the most current API version of the Ensembl Variant
  Effect Predictor.
}

\details{
  \code{VEPFlags} objects store the runtime options for querying the Ensembl
  Variant Effect Predictor (VEP). This page describes only the most current
  runtime options and is a condensed version of what is listed on the
  Ensembl web site:

  \url{http://uswest.ensembl.org/info/docs/tools/vep/script/vep_options.html}

  Runtime options for archived versions can be found on the corresponding
  archive page.

  \url{http://useast.ensembl.org/info/website/archives/index.html}
}

\author{Lori Shepherd}

\seealso{
  \itemize{
    \item The \code{ensemblVEP} function man page.
    \item The \code{VEPFlags} class man page.
  }
}

\examples{
  ## See ?VEPFlags for examples of constructing instances of a
  ## VEPFlags object with different runtime options.
}

\keyword{utilities}
