\name{VEPToGRanges}
\alias{VEPToGRanges}

%% NOTE: This function is from VariantTools and not yet integrated
%%       into REnsemblVEP 

\title{
  Convert a Variant Effect Predictor tab separated file to a GRanges
  object. 
}

\description{
 This simple function will convert a data.frame generated from
 reading in the results of running Variant Effect Predictor into
 a GRanges object with some of the associated information attached as
 ElementMetaData columns 
}

\usage{
VEPToGRanges(VEP_tab)
}

\arguments{
  \item{VEP_tab}{
    A data,frame object with appropriate column names generated by 
    reading in the output from Variant Effect Predictor
  }
}

\value{
  A GRanges object for each variant listed in the output of Variant 
  Effect Predictor. The Granges will have additional ElementMetaData 
  columns Allele (the variant allele), Gene (the gene ID given by VEP), 
  TXID (transcript ID filled in from the 'Feature' column in VEP),
  Feature_type, Consequence and Existing_variation (giving rsID or cosmic
  ID if they exist).
}

\author{
Jeremiah Degenhardt
}

\keyword{internal}
